Renamed PolicyStore file
