+++ mlte/store/user/underlying/fs.py
@@ -13,7 +13,7 @@ from mlte.store.user.mappers import (
     UserMapper,
 )
 from mlte.store.user.policy import user_policy
-from mlte.store.user.policy.policy_store import PolicyStoreService
+from mlte.store.user.policy.policy_store_service import PolicyStoreService
 from mlte.store.user.store import UserStore
 from mlte.store.user.store_session import UserStoreSession
 from mlte.user.model import (
